sparse-checkout: update locking pattern
There is a narrow window for race conditions while one process
updates the sparse-checkout file and another process tries to
do a checkout.

To avoid these races, follow this pattern in 'git sparse-checkout
set':

  1. Take index.lock.
  2. Take sparse-checkout.lock.
  3. Update the working directory according to in-memory patterns.
  4. Write patterns to sparse-checkout.lock.
  5. Commit sparse-checkout.lock.
  6. Commit index.lock.

This avoids the race condition, as any checkout must first take
the index.lock.

The code is organized into a new "struct update_data" so we can
hold the multiple structs and options across this time period.

Helped-by: Kevin Willford <kewillford@microsoft.com>
Signed-off-by: Derrick Stolee <dstolee@microsoft.com>