Documentation: add docs around testing installers
The way we test pull requests in the microsoft/git fork is a bit
strange. We rely on the integration between the Git code and a C#
layer in another repository. The functional tests in the C# layer
are interesting and sometimes the only way to fully test the
complex interactions between the components.

Document the process for generating and using an installer.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>