diff: skip batch object download when possible
When computing changed-path Bloom filters or performing a name-only
diff, we do not need the blob contents before completing the diff
values. Thus, we do not need to download a pack containing the blobs
we do not have on-disk before completing our diff calculation.

This prevents downloading every blob in a partial clone during
"git log --raw" and "git diff --name-only" commands.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>